requirements.txt
# Core dependencies
langchain-core>=0.1.14
langchain-openai>=0.0.5
langchain-community>=0.0.16
langchain-text-splitters>=0.0.1
openai>=1.3.0
streamlit>=1.27.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0

# Vector database
chromadb>=0.4.13

# Data processing
python-dotenv>=1.0.0
pydantic>=2.3.0

# Data extraction
scrapy>=2.10.0
